MDEV-23538: Rename mariadb.pc to mariadb-server-embedded.pc to avoid confusion
Change so that the client library config can be accessed using the name
mariadb.pc (or libmariadb.pc).

Rename old mariadb.pc to mariadb-embedded-server.pc to better indicate
that the config is for the daemon build, not client.

Also keep/include names with 'lib' as symlinks, as it aligns with how
most of *.pc files are named in general.

After this change we would have for the files:
- mariadb.pc for client (with libmariadb.pc as symlink)
- mariadb-embedded-sever.pc for server (with libmariadbd.pc as symlink)

(cherry picked from commit ddfebd48f58678ff31427f7f7d971981e2d098ec)