Use archive.mariadb.org as official watch source
The official source for new releases (and all the older release) is
archive.mariadb.org. This service is under MariaDB Foundation control
and monitoring and it is guaranteed to be stable and accurate in time.

(cherry picked from commit 89b4712eabb33b9ef3ef6ff3ea17a3c15e889ef8)