Salsa CI: Add arm64, armhf and riscv64 builds
As an arm64 runner is now available in the Salsa group 'mariadb-team'
start building on 'arm64' and 'armhf' architectures. Don't build on
'armel' as it fails to compile on the Salsa CI runner (potentially due
to cross-build issues).

Also build on riscv64 which has a runner available since August 2024.

Use custom "if" rules to ensure forks of this project are not affected
and will not get stuck jobs, as forks are unlikely to have custom
'arm64' tagged runners.