Deb: Adapt Salsa-CI for MariaDB 10.9+ series
After merge from 10.6 to 10.9 the references to buster-backports made
the Salsa-CI invalid. The MariaDB 10.9+ series depend in libfmt, which
isn't available for Debian 10 "Buster", so no buster-backports can be built
and thus it can't be used in Salsa-CI upgrade tests. Simply document this
fact and clean away use of buster-backports.