Run full mariadb-test-run with also big tests post build
The autopkgtests already runs with --big-test. Use this now also in the
post-build mariadb-test-run to get more exposure to how stable the big
tests are as the post-build test run occurs on so many architectures
across the Debian and Ubuntu build systems (many more than what their
CI systems run on).