Add more transitional mariadb-server-10.{1,3,6} packages for smooth upgrade
The package mariadb-server-10.5 already exists to ensure smooth upgrade
to versionless package names paradigm as described in detail in
https://bugs.debian.org/1035949.

With the same logic, it would be beneficial to also have the older MariaDB
versions that once were present in Debian/Ubuntu. This is urgent in
particular to help MariaDB 10.6 upgrades from Ubuntu 22.04 to 24.04.