Add CMake flag to ignore libfmt exit code so cross-building works
It is also safe to use `-DHAVE_SYSTEM_LIBFMT_EXITCODE=0` with < 10
version of libfmt, and in Debian there is currently 9.1.0.

It is also safe to have this flag when cross-compiling against a
libfmt-10+ version that has
https://github.com/fmtlib/fmt/commit/44c3fe1ebb466ab5c296e1a1a6991c7c7b51b72e

However, there hasn't been a libfmt release with this fix yet, so we
need to be careful to avoid libfmt 10.0.0-10.2.1.

ref:
- https://github.com/fmtlib/fmt/issues/3835
- https://alioth-lists.debian.net/pipermail/pkg-mysql-maint/2024-May/017325.html