Apply `wrap-and-sort -ast`
No functional changes.

This follows what the consensus of future defaults in Debian#895570
seems to gravitate towards, as well as aligns with the Debputy 'black'.

-a, --wrap-always
    Wrap all package lists in the Debian control file even if they do
    not exceed the line length limit and could fit in one line.

-s, --short-indent
    Only indent wrapped lines by one space (default is in-line with the
    field name).

-t, --trailing-comma
    Add a trailing comma at the end of the sorted fields. This minimizes
    future differences in the VCS commits when additional dependencies
    are appended or removed.

Also update README and CI to enforce this new style.

Specifically do not use options -b and -k yet, as it would reorder the
packages and make diffs across branches too hard to read for the time
being. Also for MariaDB it might make sense to have the main packages
listed first, and plugins last (even though that is not the alphabetical
order).

Manually exclude putting 'mariadb-plugin-rocksdb' last in the list as
it would make autopkgtests fail due to missing RocksDB plugin.