Replace autopkgtest smoke test dependency hack with arch list
Instead of expecting the system to always have no prior 'mariadb-server'
package installed and apt resolving to install 'mariadb-provider-rocksdb',
just make the package an explicit dependency and limit it to the same
list of architectures that the RocksDB plugin is being built to.

This also allows package to be fully sorted with `wrap-and-sort -ast`.