Update client program 'mariadb' trace to match new libmariadb v3.3
In upstream commit 30d3cfad the MariaDB Connector C was updated to version
v3.3, and as a result also the client program seems to now output the help
text sorted in a different way than earlier. The only new command-line
parameter is `--sandbox`, otherwise nothing was added or removed, only
listed in a different order.