Extend gbp.conf and update maintainer READMEs
Configure git-buildpackage permanently via a gbp.conf file to allow running
git-buildpackage commands without having to remember all the options.

Comment the config file extensively to explain each option. Also update
the maintainer documentation to make it explicit on what are the correct
commands to run for imports and contributions.