Make all MariaDB client plugins available in server plugin path
The MariaDB server has only one plugin directory (--plugin-dir). If the
server is running as a replica, it will act as a client when connecting
to the primary database instance. In those cases it may need to use the
client plugins, but it is unable to load the client plugins in the same
way as other client library using code would, and instead expects to find
the plugins in the server plugin directory. Add the client plugins as
symlinks there to prevent the server from failing to connect as client
when using plugin dependant client features.

This was copied by upstream in 57629c2 as part of MDEV-34744.