Update Lintian overrides for libmariadb-dev regarding man page syntax
With MariaDB 11.4.4 the MariaDB Connector C in libmariadb/ was updated
from version 3.4.1 to 3.4.2, where the manpages are generated with a new
version of Pandoc, resulting in different syntax and different errors.

The man pages still render OK to users, to the warnings can be ignored
for now, but they should be fixed upstream eventually.