Salsa CI: Stop maintaining mariadb.org-10.{2-9} upgrade tests
These tests are currently failing as they use Debian unstable, which has
new changes that would need the old tests to be adapted, and is not worth
the effort. The actual upgrades are likely to still work, and the tests
that upgrade from 10.10 and 10.11 will still remain and provide enough
coverage to detect regressions.