Clean away duplicate wsrep-provider from configuration traces
Starting with MariaDB 11.4.4 it seems that the `wsrep-provider` exists
twice as a system variable: both to control if plugin is on or off, and
to define the path to the plugin.

Remove the on/off version from traces as the order was random and led to
traces failing to match on armhf, armel and i386.

Example:
https://launchpadlibrarian.net/769497015/buildlog_ubuntu-plucky-armhf.mariadb_1%3A11.4.4-3~bpo25.04.1~1736406650.12f2b164879+dev.otto_BUILDING.txt.gz

@@ -3160,8 +3155,8 @@
 wsrep-notify-cmd
 wsrep-on                                                     FALSE
 wsrep-patch-version                                          wsrep_26.22
-wsrep-provider                                               none
 wsrep-provider                                               ON
+wsrep-provider                                               none
 wsrep-provider-options
 wsrep-recover                                                FALSE
 wsrep-reject-queries                                         NONE

Also normalize max value for quick-max-column-width to be capped at what
32-bit systems support.

(cherry picked from commit 7a432512d3f2865e00035e307029b6babdc352b2)