Salsa CI: Remove the debian/rules syntax check as not very useful
While `make` can detect simple syntax issues, the build itself is likely
to detect them too, so this test isn't very useful. Also, with the intro-
duction of dh-package-notes, this test would need a new dependency. So,
instead of bloating this job, just remove it instead.

This fixes builds failing on error:

    dh: error: Unknown sequence /usr/share/debhelper/dh_package_notes/
      package-notes.mk (choose from: binary binary-arch binary-indep build
      build-arch build-indep clean install install-arch install-indep)
    debian/rules:14: /usr/share/debhelper/dh_package_notes/
      package-notes.mk: No such file or directory