Salsa CI: Adapt piuparts helper script to new source format in Bookworm
The piuparts job started failing on `grep` not finding any apt source
files with old format:

    ERROR: Command failed (status=1): ['chroot', '/tmp/tmp27wvmxr9', 'eatmydata', 'tmp/scripts/pre_install_salsa_ci']
    + set -e
    + grep -r ^deb  /etc/apt/sources.list.d

Fix this by making the debug command more generic and start failing on
errors only after it. Also add output to make debugging and root-causing
similar failures in this script easier in the future.