Document creating app user and allowing remote connections
As MariaDB ships only with 'root'@'localhost' and 'mysql'@'localhost' by
default in Debian, users most likely need to create a separate user for
their app, and if the app is on a separate host, also allow MariaDB to
listen for connections on the network and have TLS enabled (and required).

Document these typical things to save users time in setting up their
MariaDB instance.