Make COMPILATION_COMMENT and MYSQL_SERVER_SUFFIX more useful
Having 'Debian n/a' in the version string for all binaries built in Sid
does not look good. Just state "from $VENDOR", which will be "Debian"
or "Ubuntu". The rest of the version string will be enough to see what
the exact version and Debian or Ubuntu release it is about. Note that
if binary hasn't been rebuilt in a stable update, it would still
continue to reflect Debian (Sid = N/A).

Put the $VENDOR information in server suffix as it is a more natural
place to denote which exact version/revision/build is running.

This frees up the compilation comment field to be used for important
notices to users instead.

For the sake of documentation, list of variables and example values
below:

	DEB_VERSION: 1:11.8.1-5
	DEB_VERSION_REVISION: 5
	DEB_VERSION_VERSION: 11.8.1
	DEB_VERSION_MAJOR: 11.8
	RELEASE: 'n/a' for Sid, or e.g. '25.10' for Ubuntu devel
	DEB_VENDOR: 'Debian' or 'Ubuntu'
	DEB_PARENT_VENDOR: empty in Debian but 'Debian' in Ubuntu
	DEB_SOURCE_PACKAGE_VERSION: empty
	DEB_DISTRIBUTION: UNRELEASED (or whatever d/changelog has)
	DEB_TIMESTAMP: 1745452092