Suppress native AIO warning introduced in upstream a87bb96
This is necessary to ensure among others ppc64el and s390x continues to
build successfully and will not fail the post-build mariadb-test-test
run on warnings such as:

    [Warning] InnoDB: native AIO failed: falling back to
    innodb_use_native_aio=OFF

Unlike upstream MariaDB, in Debian we removed libaio in for Linux on
MariaDB 10.6+ in 612630c6 and completely in 1d648d6f. Thus the build
will not have libaio at all on any platform in Debian, and the warning
is useless.

(cherry picked from commit 2ef00f5e5b7bb5c1631c5015e492dccf50e3ea26)