Salsa CI: Remove Buster upgrades and ignore missing Trixie ones
With Trixie released, Buster is no longer available, so remove tests that
depended on it. Also remove upgrade tests from MariaDB.org 10.5 and 11.7
as they will never be available for Trixie. The 10.6, 10.11 and 11.4 should
become available for Trixie in next release, but is not yet, so allow them
to fail for now.