Update Lintian overrides for new files in latest upstream version
This is necessary to silence new Lintian errors so that other errors
won't be masked. The use of `sphinx_rtd_theme` is now used in
Columnstore when it previously was libmarias3 only and thus needs to be
overridden for both paths.

(cherry-picked from 1b8852b439ea397b7515d6d416208c8b434a255f)