d/mariadb-common.postinst: do not require /etc/mysql/mariadb.cnf
Currently, package upgrades fail if /etc/mysql/mariadb.cnf does not
exist. The fallback update-alternatives call is allowed to fail, but
the call(s) done by /usr/share/mysql-common/configure-symlinks are not
allowed to fail.

/etc/mysql/mariadb.cnf is a conffile, hence not automatically restored
on package upgrade if missing. It is only one possible alternative for
/etc/mysql/my.cnf, hence admins might have intentionally chosen a
different one.

In any case, it should be consistent. If /etc/mysql/mariadb.cnf is
assumed to be strictly required, it must not be a conffile or actively
restored if missing.

This commit exits the postinst script positively early, if
/etc/mysql/mariadb.cnf does not exist, but prints a notice about that.
The explicit "|| true" which allowed update-alternatives to fail is
removed: it is now skipped as well if /etc/mysql/mariadb.cnf does not
exist, but causes the script to error out in case of other failures.