Disable cracklib plugin on hurd-amd64 (Closes: #1122328)
The package `cracklib-runtime` is not available on `hurd-amd64`. Work
around this so that builds on `hurd-amd64` can start, and we can see if
they potentially fail on something else (potentially similar issues as
in `hurd-i386`). The cracklib plugin is not essential for MariaDB, so
this is an acceptable drawback for Hurd users.