Make mariadb-test-run logs more informative
Remove the --verbose-start option as it was way too verbose and just
cluttered the logs.

Increase the default --max-test-fail from 10 to 30, as it seems that
having a dozen or so failures is common, and it is better to continue
the run to the end and then be able to analyze all errors from a single
log instead of focusing on fixing only the fist failures.