Salsa CI: Remove failing 'mysql.com-5.7 with Buster upgrade' job
The job 'mysql.com-5.7 with Buster upgrade' last passed on December 6th,
2025, and was failing on a run a week later. The error log ends on apt
nagging about "unmet dependencies", but comparing the logs of the two runs
there is nothing that changed. Most likely the support for MySQL 5.7 or
Buster in general was partially deprecated on the repo.mysql.com server
without fully shutting down the repository.

Instead of sinking effort in debugging this further, simply disable the
job as it failing is clearly not a sign on anything regressing in Debian
(as the same commit with no Debian changes is now failing) and keeping
tests for upgrading from Buster/MySQL 5.7 has very low value now as both
of them are long deprecated.

deprecated.